Extract new queryDcs module for fetching data from DCS
